+++ commons_android/src/test/java/com/github/barriosnahuel/vossosunboton/commons/android/analytics/AnalyticsCoverageMatrixTest.kt
@@ -65,6 +65,8 @@ internal class AnalyticsCoverageMatrixTest {
                 "AboutCreditsOpen",
                 "AboutLicenseOpen",
                 "AboutSourceOpen",
+                "AboutPrivacyPolicyOpen",
+                "AboutDataSafetyOpen",
                 "AboutBrandingAudioPlay",
                 "MilestoneAudios",
             )